+++ sdk/typescript/src/types.ts
@@ -1,4 +1,4 @@
-export const VERSION = "0.6.5";
+export const VERSION = "0.6.6";
 export const PROTOCOL = 14;
 export const REQUIRES_CLI = "0.43.4";